refactor(main): extract visible-overlay platform interaction runtime from main.ts
